requirements.txt
spotipy>=2.25.0
requests>=2.32.3
beautifulsoup4>=4.12.3
pytz>=2024.1
python-dateutil>=2.8.2
python-dotenv>=1.0.0 